+++ packages/backend/src/plugins/scaffolder.ts
@@ -33,7 +33,7 @@ import {
   createAppendFileAction,
   createMergeJSONAction,
 } from '@roadiehq/scaffolder-backend-module-utils';
-import { createAwsS3CpAction } from '@roadiehq/scaffolder-backend-module-aws';
+import { createAwsS3CpAction, createEcrAction } from '@roadiehq/scaffolder-backend-module-aws';
 import Docker from 'dockerode';
 import { Router } from 'express';
 import type { PluginEnvironment } from '../types';
@@ -62,6 +62,7 @@ export const createActions = (options: {
     createAppendFileAction(),
     createMergeJSONAction({}),
     createAwsS3CpAction(),
+    createEcrAction(),
     createHttpBackstageAction({ config }),
     ...defaultActions,
   ];